Crash prevention and bugs fixes
All of the following issues were resolved and tested.

Issue #1: requestpassword() immediately showed error message after asking for input, regardless of the input.

Issue #2: In requestPassword(), it was possible to answer "no" to all questions and exclude all types of characters, which leaded to a crash when the password was generated.

Issue #3: PasswordRequestError() would inform the user that the input was invalid but did not ask for new input, leading to undesired results.